+++ Optimizer/Controls/AppCard.cs
@@ -1,5 +1,4 @@
 ﻿using System.Windows.Forms;
-using System.Diagnostics;
 
 namespace Optimizer
 {